MyApplication/app/chatbot/requirements.txt
flask
mysql-connector-python
sentence-transformers
faiss-cpu
numpy
scikit-learn
